requirements.txt
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
